+++ src/main/java/seedu/address/logic/parser/AddCommandParser.java
@@ -15,7 +15,6 @@ import java.util.stream.Stream;
 
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.logic.commands.AddCommand;
-import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.customfields.CustomField;
 import seedu.address.model.group.Group;
