+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -50,61 +50,61 @@ public class AddressBookParser {
         final String arguments = matcher.group("arguments");
         switch (commandWord) {
 
-            case AddCommand.COMMAND_WORD:
-            case AddCommand.COMMAND_ALIAS:
-                return new AddCommandParser().parse(arguments);
+        case AddCommand.COMMAND_WORD:
+        case AddCommand.COMMAND_ALIAS:
+            return new AddCommandParser().parse(arguments);
 
-            case EditCommand.COMMAND_WORD:
-            case EditCommand.COMMAND_ALIAS:
-                return new EditCommandParser().parse(arguments);
+        case EditCommand.COMMAND_WORD:
+        case EditCommand.COMMAND_ALIAS:
+            return new EditCommandParser().parse(arguments);
 
-            case SelectCommand.COMMAND_WORD:
-            case SelectCommand.COMMAND_ALIAS:
-                return new SelectCommandParser().parse(arguments);
+        case SelectCommand.COMMAND_WORD:
+        case SelectCommand.COMMAND_ALIAS:
+            return new SelectCommandParser().parse(arguments);
 
-            case DeleteCommand.COMMAND_WORD:
-            case DeleteCommand.COMMAND_ALIAS:
-                return new DeleteCommandParser().parse(arguments);
+        case DeleteCommand.COMMAND_WORD:
+        case DeleteCommand.COMMAND_ALIAS:
+            return new DeleteCommandParser().parse(arguments);
 
-            case ClearCommand.COMMAND_WORD:
-            case ClearCommand.COMMAND_ALIAS:
-                return new ClearCommand();
+        case ClearCommand.COMMAND_WORD:
+        case ClearCommand.COMMAND_ALIAS:
+            return new ClearCommand();
 
-            case FindCommand.COMMAND_WORD:
-            case FindCommand.COMMAND_ALIAS:
-                return new FindCommandParser().parse(arguments);
+        case FindCommand.COMMAND_WORD:
+        case FindCommand.COMMAND_ALIAS:
+            return new FindCommandParser().parse(arguments);
 
-            case ListCommand.COMMAND_WORD:
-            case ListCommand.COMMAND_ALIAS:
-                return new ListCommandParser().parse(arguments);
+        case ListCommand.COMMAND_WORD:
+        case ListCommand.COMMAND_ALIAS:
+            return new ListCommandParser().parse(arguments);
 
-            case SortCommand.COMMAND_WORD:
-            case SortCommand.COMMAND_ALIAS:
-                return new SortCommand();
+        case SortCommand.COMMAND_WORD:
+        case SortCommand.COMMAND_ALIAS:
+            return new SortCommand();
 
-            case HistoryCommand.COMMAND_WORD:
-            case HistoryCommand.COMMAND_ALIAS:
-                return new HistoryCommand();
+        case HistoryCommand.COMMAND_WORD:
+        case HistoryCommand.COMMAND_ALIAS:
+            return new HistoryCommand();
 
-            case ExitCommand.COMMAND_WORD:
-                return new ExitCommand();
+        case ExitCommand.COMMAND_WORD:
+            return new ExitCommand();
 
-            case HelpCommand.COMMAND_WORD:
-                return new HelpCommand();
+        case HelpCommand.COMMAND_WORD:
+            return new HelpCommand();
 
-            case UndoCommand.COMMAND_WORD:
-            case UndoCommand.COMMAND_ALIAS:
-                return new UndoCommand();
+        case UndoCommand.COMMAND_WORD:
+        case UndoCommand.COMMAND_ALIAS:
+            return new UndoCommand();
 
-            case RedoCommand.COMMAND_WORD:
-            case RedoCommand.COMMAND_ALIAS:
-                return new RedoCommand();
+        case RedoCommand.COMMAND_WORD:
+        case RedoCommand.COMMAND_ALIAS:
+            return new RedoCommand();
 
-            case PasswordCommand.COMMAND_WORD:
-            case PasswordCommand.COMMAND_ALIAS:
-                return new PasswordCommandParser().parse(arguments);
-            default:
-                throw new ParseException(MESSAGE_UNKNOWN_COMMAND);
+        case PasswordCommand.COMMAND_WORD:
+        case PasswordCommand.COMMAND_ALIAS:
+            return new PasswordCommandParser().parse(arguments);
+        default:
+            throw new ParseException(MESSAGE_UNKNOWN_COMMAND);
         }
     }
 }
